FIR IDE: extract PsiType conversions to KtPsiTypeProvider
