+++ idea/idea-frontend-api/src/org/jetbrains/kotlin/idea/frontend/api/components/KtTypeProvider.kt
@@ -5,13 +5,10 @@
 
 package org.jetbrains.kotlin.idea.frontend.api.components
 
-import com.intellij.psi.PsiType
 import org.jetbrains.kotlin.idea.frontend.api.ValidityTokenOwner
 import org.jetbrains.kotlin.idea.frontend.api.symbols.KtNamedClassOrObjectSymbol
 import org.jetbrains.kotlin.idea.frontend.api.types.KtType
 import org.jetbrains.kotlin.idea.frontend.api.types.KtTypeNullability
-import org.jetbrains.kotlin.load.kotlin.TypeMappingMode
-import org.jetbrains.kotlin.psi.KtTypeReference
 
 abstract class KtTypeProvider : KtAnalysisSessionComponent() {
     abstract val builtinTypes: KtBuiltinTypes
@@ -19,8 +16,6 @@ abstract class KtTypeProvider : KtAnalysisSessionComponent() {
     abstract fun approximateToSuperPublicDenotableType(type: KtType): KtType?
 
     abstract fun buildSelfClassType(symbol: KtNamedClassOrObjectSymbol): KtType
-
-    abstract fun getPsiType(ktTypeReference: KtTypeReference, mode: TypeMappingMode): PsiType
 
     abstract fun withNullability(type: KtType, newNullability: KtTypeNullability): KtType
 }
@@ -40,9 +35,6 @@ interface KtTypeProviderMixIn : KtAnalysisSessionMixIn {
 
     fun KtNamedClassOrObjectSymbol.buildSelfClassType(): KtType =
         analysisSession.typeProvider.buildSelfClassType(this)
-
-    fun KtTypeReference.getPsiType(mode: TypeMappingMode = TypeMappingMode.DEFAULT): PsiType =
-        analysisSession.typeProvider.getPsiType(this, mode)
 
     fun KtType.withNullability(newNullability: KtTypeNullability): KtType =
         analysisSession.typeProvider.withNullability(this, newNullability)
