+++ idea/idea-frontend-fir/src/org/jetbrains/kotlin/idea/frontend/api/fir/components/KtFirExpressionTypeProvider.kt
@@ -6,30 +6,23 @@
 package org.jetbrains.kotlin.idea.frontend.api.fir.components
 
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiType
 import org.jetbrains.kotlin.fir.declarations.FirCallableDeclaration
 import org.jetbrains.kotlin.fir.expressions.*
 import org.jetbrains.kotlin.fir.psi
-import org.jetbrains.kotlin.fir.references.FirErrorNamedReference
 import org.jetbrains.kotlin.fir.references.FirNamedReference
-import org.jetbrains.kotlin.fir.references.FirResolvedNamedReference
-import org.jetbrains.kotlin.fir.resolve.diagnostics.ConeUnresolvedNameError
 import org.jetbrains.kotlin.fir.typeContext
-import org.jetbrains.kotlin.fir.types.ConeClassErrorType
-import org.jetbrains.kotlin.fir.types.ConeKotlinType
 import org.jetbrains.kotlin.fir.types.coneType
 import org.jetbrains.kotlin.idea.fir.low.level.api.api.getOrBuildFir
 import org.jetbrains.kotlin.idea.fir.low.level.api.api.getOrBuildFirOfType
 import org.jetbrains.kotlin.idea.fir.low.level.api.api.getOrBuildFirSafe
 import org.jetbrains.kotlin.idea.frontend.api.components.KtExpressionTypeProvider
 import org.jetbrains.kotlin.idea.frontend.api.fir.KtFirAnalysisSession
+import org.jetbrains.kotlin.idea.frontend.api.fir.utils.getReferencedElementType
 import org.jetbrains.kotlin.idea.frontend.api.tokens.ValidityToken
 import org.jetbrains.kotlin.idea.frontend.api.types.KtErrorType
 import org.jetbrains.kotlin.idea.frontend.api.types.KtType
 import org.jetbrains.kotlin.idea.frontend.api.withValidityAssertion
-import org.jetbrains.kotlin.idea.references.FirReferenceResolveHelper
 import org.jetbrains.kotlin.lexer.KtTokens
-import org.jetbrains.kotlin.load.kotlin.TypeMappingMode
 import org.jetbrains.kotlin.psi.*
 
 internal class KtFirExpressionTypeProvider(
@@ -48,29 +41,6 @@ internal class KtFirExpressionTypeProvider(
             is FirStatement -> with(analysisSession) { builtinTypes.UNIT }
             else -> error("Unexpected ${fir::class}")
         }
-    }
-
-    override fun getPsiTypeForKtExpression(
-        expression: KtExpression,
-        mode: TypeMappingMode,
-    ): PsiType = withValidityAssertion {
-        when (val fir = expression.getOrBuildFir(firResolveState)) {
-            is FirExpression -> fir.typeRef.coneType.asPsiType(mode, expression)
-            is FirNamedReference -> fir.getReferencedElementType().asPsiType(mode, expression)
-            is FirStatement -> PsiType.VOID
-            else -> error("Unexpected ${fir::class}")
-        }
-    }
-
-    private fun FirNamedReference.getReferencedElementType(): ConeKotlinType {
-        val symbols = when (this) {
-            is FirResolvedNamedReference -> listOf(resolvedSymbol)
-            is FirErrorNamedReference -> FirReferenceResolveHelper.getFirSymbolsByErrorNamedReference(this)
-            else -> error("Unexpected ${this::class}")
-        }
-        val firCallableDeclaration = symbols.singleOrNull()?.fir as? FirCallableDeclaration<*>
-        return firCallableDeclaration?.returnTypeRef?.coneType
-            ?: ConeClassErrorType(ConeUnresolvedNameError(name))
     }
 
     override fun getExpectedType(expression: PsiElement): KtType? {
